Делит код на модули по заданию
